Allow running pyro_sim.py from anywhere
I changed setup.py to install an autogenerated shim script that loads
pyro.pyro_sim as a module, as otherwise `__file__` would point to the
bin directory it gets installed to, rather than the pyro/ directory.

With a minor change to test.py, this lets us run the regression tests
from outside pyro/ as well.