Fix more pylint warnings (#143)
Fix a spurious divide-by-zero error in sedov_compare.py and
gauss_diffusion_compare.py that pylint found. Most of the other fixes
are for `unused-variable`, plus some instances of
`consider-using-enumerate`, `consider-using-dict-items`,
`consider-iterating-dictionary`, `use-dict-literal`,
`useless-object-inheritance`, and `unneeded-not`.

Also disable `pointless-string-statement`, since these are used for
block comments in a couple places.