convergence utility tools (#165)
This pr makes convergence.py more general so that it takes any variable we want with the arbitrary multiplicative factor of resolutions apart between plot files. The default is still density and twice the resolution between the fine plot files and the coarse plot files.

convergence_plot.py allows users to give 3 or more output files which the corresponding resolution differs by a constant multiplicative constant.

The goal here is to duplicate the Richardson utility tool that we had in castro/amrex. It outputs the table that tells us the convergence rate as well as a plot.