Create incompressible_viscous solver and lid-driven cavity test problem (#138)
Needed to change a few things in incompressible/simulation.py as well:

Neumann BC's on phi in the case of dirichlet BC's on v (and apply phi-specific BC's in the multigrid solves)
Allow user-defined BC's
Allow another method for the velocity update in the evolve loop
The latter two are then redefined in incompressible_viscous/simulation.py. The rest of the methods are inherited from the base class.

For the lid-driven cavity problem, the only free parameter is the viscosity. Since the velocity and length scales are =1, the Reynolds number is just 1/viscosity.