add some protection against negatives in 4th order (#309)
this now uses a mask when we convert from averages to centers.  If either density or internal
energy become negative, then we flag the mask for that zone.  We then go back and reset all of the flagged zones to the cell averages -- this will drop the method to 2nd order in that zone, but its better than having negative density because of a steep gradient.